data/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpraw==7.7.1
python-dotenv==1.0.0
pymongo==4.6.0
requests==2.31.0
